+++ tests/MiniZinc.SolveTests/ExampleTests.fs
@@ -1,8 +1,0 @@
-namespace MiniZinc.Solve.Tests
-
-open MiniZinc
-open Xunit
-open System.IO
-
-module ExampleTests =
-    ()